ACC-2208 fix sonar issues

+++ contentgrid-appserver-domain/src/main/java/com/contentgrid/appserver/domain/DatamodelApiImpl.java
@@ -145,10 +145,10 @@ public class DatamodelApiImpl implements DatamodelApi {
         }
     }
 
-    private ItemCount calculateCount(Supplier<ItemCount> countSupplier, OffsetData offsetData, int size, boolean hasNext) {
+    private ItemCount calculateCount(Supplier<ItemCount> countSupplier, OffsetData offsetData, long size, boolean hasNext) {
         var hasPrevious = offsetData.getOffset() > 0;
 
-        if (!hasNext && !(hasPrevious && size == 0)) {
+        if (!hasNext && !(hasPrevious && size == 0L)) {
             // If this is exactly the last page with results: we know the exact size, no need for counting
             return ItemCount.exact(offsetData.getOffset() + size);
         }
@@ -157,7 +157,7 @@ public class DatamodelApiImpl implements DatamodelApi {
 
         if (hasNext) {
             // There has to be a next page, adjust count to have at least one item on the next page
-            return result.orMinimally(offsetData.getOffset() + offsetData.getLimit() + 1);
+            return result.orMinimally(offsetData.getOffset() + offsetData.getLimit() + 1L);
         } else {
             // There is no next page and there are also no results on this page (otherwise we returned exact result),
             // adjust count to have at most the amount on the previous page
